fix(testing): added windows tests (#25)
## changes

- [x] added windows tests
- [] ~~allow overriding of environment file with config file~~


## why?

[no longer valid: not doing this]

i've changed the order of operations so that config file overrides
environment variables. the reason for this is because on compute
benches, environment variables are baked in and never change once
jupyter lab starts. so once a bench is running, it is not possible to
change env variables that are read by the client. and because env
variables override files (previously), it was impossible to change any
settings for the client.