feat: disabled printing context for compute cluster and hardware which the platform no longer provides (#53)
`deeporigin context` is intended to print out information including
- The compute cluster where the bench is being run. This is to help
developers, such as molecular modeling, communicate with other
containers.
- Hardware. This is to provide users, such as molecular modeling,
accurate information about the resources of the bench. This is important
because the Linux commands for reporting memory don't work as expected
in containers in shared hosts.

This PR disables printing this information because the platform no
longer provides this information.
[DOS-3008](https://deeporiginbio.atlassian.net/browse/DOS-3008) will add
it. Until that point, we should avoid printing unsupported information.

Rather than deleting support for this information in this library, I
instead refactored the library to make the printing of this information
conditional. Once the platform resumes providing this information, we
can simply change the default values of these conditions.

The second commit of this PR also formats a couple of source files.

The third commit of this PR refactors the generation of the string
representation of context from the CLI module to `__str__` methods for
the classes which capture context information. This was done to make it
easier to pass conditional information for testing.

Closes [DOS-3009](https://deeporiginbio.atlassian.net/browse/DOS-3009)