fix: more bugfixes & improvements (#83)
## changes

- [x] wrapped all pretty printing code for data frames in a try so that
this never breaks important code
- [x] overrides for head and tail methods
- [x] better visibility in dark mode for dataframe headers
- [ ] ability to retrieve notebook entries 
- [x] ability to add a new column in a dataframe
- [x] caching of user names for faster displays

## technical discussion 

previously this behavior existed, where printing a dataframe worked, but
using df.head() would trigger a spurious warning about dataframes
containing modified data

<img width="758" alt="dataframe-warnings 9 07 43 AM"
src="https://github.com/user-attachments/assets/64b29de3-cca5-4b28-a92c-78d5d70eb1cc">

this was because this was technically correct -- pandas df.head()
created a shallow copy of a dataframe, which was technically a new
object, and because the data in the new dataframe was not the same, the
warning was shown.

however, this is useless and misleading from an end user persp. so we
override these methods.