chore: added src/README describing client code structure, comments to chemistry.py (#168)
Wrote a general overview README that describes the general structure of
code in "src/" submodules, including implementaion details for some code
components. This is a useful place for someone wanting to understand the
code and interaction between different subsystems/modules/repos.

Updated docstring to provide a nice module description in docs. Fixed
for new file location.

---------

Co-authored-by: mike7ant <16749494+mike7ant@users.noreply.github.com>