feat: support for client pattern in dd toolkit (#223)
## problem description

balto (and other machine users) need to be able to use the client in a
way that allows for multi-org, multi-user. basically, it should be
possible to make a list of calls to various functions in the drug
discovery toolkit, where each call is for a different user/different org

this is a problem because the current version is designed for a
single-user, single org paradigm. tokens, org ids and other config
information is read from a file that exists on the home dir

this cannot work for this use case

## solution

assuming you have some org_id and token, you can create an instance of
`PlatformClients` that contains all clients to interact with the
platform:


```python
clients = PlatformClients(token=..., base_url=...)
```

and this can be used in"

```python
sim = Complex.from_dir(
    EXAMPLE_DATA_DIR,
   _platform_clients=clients,
)

```

this client pattern also works for all functions, e.g.:

```
get_dataframe(_platform_clients=clients)
```