feat: overhaul of how ligands are dumped to files (#292)
## problem description

a ligand can be made from a SDF file that has more than 1 ligand. how do
we communicate this to tools that require a single ligand in a single
SDF file?

the solution is to dump state to a SDF file, and hash it, and push that
SDF file to a file named by the hash on UFA

## changes

- [x] entitity.to_hash
- [x] ligand.to_hash
- [x] protein.to_hash
- [x] tests for to_hash methods
- [ ] running abfe should use ligand object, and use the ligand's upload
method
- [ ] running abfe should use the protein object, and use the ligand's
upload method
- [ ] running rbfe should use ligand object, and use the ligand's upload
method
- [ ] running rbfe should use the protein object, and use the ligand's
upload method
- [ ] running docking should use ligand object, and use the ligand's
upload method
- [ ] running docking should use the protein object, and use the
ligand's upload method
- [x] `abfe.run_end_to_end` --> `abfe.run`
- [ ] documentation changes
- [x] `Complex.from_dir` works with a SDF file with more than 1 ligand